better SOC for socketserver
