+++ src/main/java/GameManager.java
@@ -16,8 +16,8 @@ public class GameManager {
     public void stopGame() {
         game.quit();
         game = new Game(iface);
-        iface.clearConnections();
-        server.clearConnections();
+//        iface.clearConnections();
+//        server.clearConnections();
     }
 
     public Game getCurrentGame() {
